test(dcp): relocate cat_ranges unit tests into test_core.py and test_extended_gcsfs.py
